requirements.txt
django==4.1.13
djangorestframework==3.14.0
requests==2.31.0
drf-spectacular[sidecar]==0.26.5
drf-standardized-errors[openapi]==0.12.6
black==23.11.0